Local Storage 03: Criando e utilizando exceptions personalizadas

+++ igniteteams/src/storage/group/groupCreate.ts
@@ -1,12 +1,20 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
+
 import { GROUP_COLLECTION } from "@storage/storageConfig";
-import {groupsGetAll} from "@storage/group/groupsGetAll";
+import { groupsGetAll } from "@storage/group/groupsGetAll";
+
+import { AppError } from "@utils/AppError";
 
 export const groupCreate = async (newGroup: string) => {
   try {
-    const storedGroups = await groupsGetAll()
+    const storedGroups = await groupsGetAll();
 
-    const storage = JSON.stringify([...storedGroups, newGroup])
+    const groupAlreadyExists = storedGroups.includes(newGroup);
+    if (groupAlreadyExists) {
+      throw new AppError("Já existe um grupo cadastrado com esse nome.");
+    }
+
+    const storage = JSON.stringify([...storedGroups, newGroup]);
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
   } catch(error) {
     throw error;
